Refactor: Move tool views to `project_tools/` subdirectory
The following files have been moved into a new `lib/project_tools/` directory:
- `lib/compass_tool_view.dart` -> `lib/project_tools/compass_tool_view.dart`
- `lib/map_tool_view.dart` -> `lib/project_tools/map_tool_view.dart`
- `lib/points_tool_view.dart` -> `lib/project_tools/points_tool_view.dart`

`lib/project_details_page.dart` has been updated to reflect these new import paths.

Additionally, the SnackBar message in `CompassToolView` that was shown when "Add Point" was tapped has been commented out. This is to avoid showing duplicate messages if `ProjectDetailsPage` also shows a SnackBar for the same action.

In `ProjectDetailsPage`, when a point is added from the compass tool, the view will now automatically switch to the `PointsToolView` if the compass tool was active.