Use git submodule for licensed_features_package
This commit changes the `licensed_features_package` from a direct directory within the repository to a git submodule.

The submodule is configured to point to `git@github.com:jlbbooks/teleferika_licenced_packages.git`.

The `.gitignore` file has been updated to no longer ignore the `licensed_features_package/` directory, as it's now managed as a submodule.