Implement Distance Display for Points in Point Details Page
This commit introduces a new feature in the Point Details Page and Point Details Panel that displays the distance from the previous point. The distance is calculated using the existing `distanceFromPoint` method in the PointModel. Localization strings for both English and Italian have been added to support this new feature, enhancing the user experience by providing contextual distance information. Additionally, minor cleanup was performed in the project model file.